General: Moved some widgets
