+++ lib/ui/views/docs_pages/minitel_doc.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:minitel_toolbox/core/constants/app_constants.dart';
 import 'package:minitel_toolbox/ui/shared/text_styles.dart';
-import 'package:minitel_toolbox/ui/widgets/cards.dart';
+import 'package:minitel_toolbox/ui/widgets/docs_widgets.dart';
 
 class MinitelDoc extends StatelessWidget {
   const MinitelDoc({
