+++ lib/ui/views/docs_pages/wiki_docs/virtual_machines_doc.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:minitel_toolbox/ui/shared/text_styles.dart';
-import 'package:minitel_toolbox/ui/widgets/cards.dart';
+import 'package:minitel_toolbox/ui/widgets/docs_widgets.dart';
 
 class VirtMachineDoc extends StatelessWidget {
   const VirtMachineDoc({
