+++ lib/ui/views/reporting_tabs/report_tab.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:minitel_toolbox/core/funcs/url_launch.dart';
 import 'package:minitel_toolbox/ui/shared/text_styles.dart';
-import 'package:minitel_toolbox/ui/widgets/cards.dart';
+import 'package:minitel_toolbox/ui/widgets/docs_widgets.dart';
 
 class ReportTab extends StatefulWidget {
   final TextEditingController titleController;
